feat(lottoStore): lotto 판매점 기능 추가
- lotto 가격정보를 lottoTicket -> lottoStore로 이동
- 가격에 맞게 로또티켓을 리턴
